requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.26.2
omegaconf==2.3.0
pytorch-lightning==2.1.2
einops==0.7.0
taming-transformers-rom1504==0.0.6
